Move OSI commands into the interactive app
