+++ interactive/management/commands/osi_release.py
