+++ interactive/management/commands/osi_run_and_release.py
