+++ pyproject.toml
@@ -38,6 +38,9 @@ omit = [
   "jobserver/asgi.py",
   "jobserver/jobs/*",
   "interactive/management/commands/create_testing_analysis.py",
+  "interactive/management/commands/osi_release.py",
+  "interactive/management/commands/osi_run.py",
+  "interactive/management/commands/osi_run_and_release.py",
   "jobserver/management/commands/count_rows.py",
   "jobserver/management/commands/ensure_admins.py",
   "jobserver/management/commands/ensure_backends.py",
@@ -45,9 +48,6 @@ omit = [
   "jobserver/management/commands/get_transitional_projects.py",
   "jobserver/management/commands/inflate_release_files.py",
   "jobserver/management/commands/list_missing_repos.py",
-  "jobserver/management/commands/osi_release.py",
-  "jobserver/management/commands/osi_run.py",
-  "jobserver/management/commands/osi_run_and_release.py",
   "jobserver/management/commands/release.py",
   "jobserver/settings.py",
   "jobserver/wsgi.py",
